api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.31
pydantic==2.7.1
pydantic-settings==2.2.1
requests==2.31.0
python-dotenv==1.0.1
python-dotenv==1.0.1

psycopg2-binary==2.9.9
